fix: asset id int -> bigint

+++ services/admin-api/src/migrations/02-blockchain/01-exchange/1563804001220-create-asset-component.ts
@@ -37,7 +37,7 @@ export class CreateAssetComponent1563804001220 implements MigrationInterface {
         },
         {
           name: "asset_id",
-          type: "int",
+          type: "bigint",
         },
       ],
       foreignKeys: [
