+++ services/admin-api/src/migrations/02-blockchain/04-mechanics/04-craft/1653616448000-create-craft.ts
@@ -22,11 +22,11 @@ export class CreateCraft1653616448000 implements MigrationInterface {
         },
         {
           name: "item_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "price_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "merchant_id",
