+++ services/admin-api/src/migrations/02-blockchain/04-mechanics/05-promo/1658980521000-create-asset-promo.ts
@@ -14,11 +14,11 @@ export class CreateAssetPromoAt1658980521000 implements MigrationInterface {
         },
         {
           name: "item_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "price_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "merchant_id",
