+++ services/admin-api/src/migrations/02-blockchain/04-mechanics/06-grade/1657846587000-create-grade.ts
@@ -42,7 +42,7 @@ export class CreateGrade1657846587000 implements MigrationInterface {
         },
         {
           name: "price_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "contract_id",
