+++ services/admin-api/src/migrations/02-blockchain/04-mechanics/10-ponzi/1660436477200-create-ponzi-rules.ts
@@ -61,11 +61,11 @@ export class CreatePonziRules1660436477200 implements MigrationInterface {
         },
         {
           name: "deposit_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "reward_id",
-          type: "int",
+          type: "bigint",
         },
         {
           name: "external_id",
